Day 2 - Done!
1.	Step 47. Having trouble on the label in the string literal.
Solution: I just follow the steps input some formatted strings.
2.	Either wrong code or syntax error. I’m trying to add a input type text.
Solution: Just added or place on a newline but still inside the string literal.
3.	targetInputContainer.insertAdjacentHTML() += HTMLString; Syntax error.
Solution: Cannot used += when calling a function and passing arguments on it.
4.	if (invalidInputMatch) { alert(Invalid Input: ${invalidInputMatch[item]}); } Using a template literal, in your if block, call the alert() function to tell the user "Invalid Input: ", followed by the first value in the invalidInputMatch array. So im trying to get the first element in an array of the list. To show the invalid element.
Solution:  [0] is the first element as always.
5.	Global error = false. But it wont work. Syntax error?
Solution: Vaguely instruction. Basically the global error refers to the isError variable. So I call it and assigned false.